job_creator/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "job_creator"
version = "0.1.0"
dependencies = [
    "click",
    "gitpython",
    "PyYAML"
]

[project.scripts]
jc = "job_creator.__main__:main"
